Move interpreters from MiniBackend into MockInterpreters
Make the interpreters indepedent of the `State MiniBackend`.
