+++ libs/wire-subsystems/test/unit/Wire/MockInterpreters.hs
@@ -4,9 +4,13 @@ module Wire.MockInterpreters (module MockInterpreters) where
 -- ls libs/wire-subsystems/test/unit/Wire/MockInterpreters | sed 's|\(.*\)\.hs|import Wire.MockInterpreters.\1 as MockInterpreters|'                                                                                                  ─╯
 
 import Wire.MockInterpreters.Error as MockInterpreters
+import Wire.MockInterpreters.GalleyAPIAccess as MockInterpreters
 import Wire.MockInterpreters.HashPassword as MockInterpreters
 import Wire.MockInterpreters.Now as MockInterpreters
 import Wire.MockInterpreters.PasswordResetCodeStore as MockInterpreters
 import Wire.MockInterpreters.PasswordStore as MockInterpreters
 import Wire.MockInterpreters.SessionStore as MockInterpreters
+import Wire.MockInterpreters.UserEvents as MockInterpreters
+import Wire.MockInterpreters.UserKeyStore as MockInterpreters
+import Wire.MockInterpreters.UserStore as MockInterpreters
 import Wire.MockInterpreters.UserSubsystem as MockInterpreters
