+++ libs/wire-subsystems/test/unit/Wire/MockInterpreters/GalleyAPIAccess.hs
@@ -1,0 +1,19 @@
+module Wire.MockInterpreters.GalleyAPIAccess where
+
+import Imports
+import Polysemy
+import Wire.API.Team.Feature
+import Wire.API.Team.Member
+import Wire.GalleyAPIAccess
+
+-- | interprets galley by statically returning the values passed
+miniGalleyAPIAccess ::
+  -- | what to return when calling GetTeamMember
+  Maybe TeamMember ->
+  -- | what to return when calling GetAllFeatureConfigsForUser
+  AllFeatureConfigs ->
+  InterpreterFor GalleyAPIAccess r
+miniGalleyAPIAccess member configs = interpret $ \case
+  GetTeamMember _ _ -> pure member
+  GetAllFeatureConfigsForUser _ -> pure configs
+  _ -> error "uninterpreted effect: GalleyAPIAccess"
