+++ libs/wire-subsystems/wire-subsystems.cabal
@@ -177,11 +177,15 @@ test-suite wire-subsystems-tests
     Wire.MiniBackend
     Wire.MockInterpreters
     Wire.MockInterpreters.Error
+    Wire.MockInterpreters.GalleyAPIAccess
     Wire.MockInterpreters.HashPassword
     Wire.MockInterpreters.Now
     Wire.MockInterpreters.PasswordResetCodeStore
     Wire.MockInterpreters.PasswordStore
     Wire.MockInterpreters.SessionStore
+    Wire.MockInterpreters.UserEvents
+    Wire.MockInterpreters.UserKeyStore
+    Wire.MockInterpreters.UserStore
     Wire.MockInterpreters.UserSubsystem
     Wire.NotificationSubsystem.InterpreterSpec
     Wire.UserStoreSpec
